requirements.txt

# Password Quest - Dependencies
# Install with: pip install -r requirements.txt

# Core Dependencies (from password-strength-checker)
# ====================

# HTTP library for API requests (HIBP breach checking)
requests>=2.28.0

# Async HTTP library for concurrent API requests
aiohttp>=3.8.0

# Password strength estimation (Dropbox's zxcvbn)
zxcvbn-python>=4.4.24

# Environment variable management
python-dotenv>=1.0.0

# YAML configuration file support
PyYAML>=6.0.0

# TOML configuration file support (Python < 3.11)
tomli>=2.0.0; python_version<"3.11"

# Data validation for configuration
pydantic>=2.0.0


# Web Application Dependencies
# ============================

# Flask web framework
flask>=3.0.0

# CORS support for API
flask-cors>=4.0.0

# Database ORM
flask-sqlalchemy>=3.1.0


# WSGI Server for production
gunicorn>=20.1.0

# Testing Dependencies
# ====================

pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-asyncio>=0.21.0
